Write event to file
Created a remove front in linked list that is used in the file dataIO.  In dataIO, the eventArray function takes in a list of events, turns it to an array, and returns the array.  This is returned to addevent and is written to file.  This happens after every new event is added